+++ packages/b2c-dx-mcp/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @salesforce/b2c-dx-mcp
 
+## 1.0.10
+
+### Patch Changes
+
+- Updated dependencies [[`ee735bb`](https://github.com/SalesforceCommerceCloud/b2c-developer-tooling/commit/ee735bb0acac89999114c80679b4766216bf463a), [`ee735bb`](https://github.com/SalesforceCommerceCloud/b2c-developer-tooling/commit/ee735bb0acac89999114c80679b4766216bf463a)]:
+  - @salesforce/b2c-tooling-sdk@1.5.0
+
 ## 1.0.9
 
 ### Patch Changes